Revert ScopeManager refactors and organization
