+++ src/Datadog.Trace.ClrProfiler.Managed/Integrations/AspNetHttpModule.cs
