+++ src/Datadog.Trace.ClrProfiler.Managed/Integrations/AspNetScopeManager.cs
@@ -1,0 +1,102 @@
+#if !NETSTANDARD2_0
+
+using System;
+using System.Web;
+using Datadog.Trace.Logging;
+
+namespace Datadog.Trace.ClrProfiler.Integrations
+{
+    internal class AspNetScopeManager : IScopeManager
+    {
+        private static readonly ILog Log = LogProvider.For<AspNetScopeManager>();
+
+        private readonly string _name = "__Datadog_Scope_Current__" + Guid.NewGuid();
+        private readonly AsyncLocalCompat<Scope> _activeScopeFallback = new AsyncLocalCompat<Scope>();
+
+        public event EventHandler<SpanEventArgs> SpanOpened;
+
+        public event EventHandler<SpanEventArgs> SpanActivated;
+
+        public event EventHandler<SpanEventArgs> SpanDeactivated;
+
+        public event EventHandler<SpanEventArgs> SpanClosed;
+
+        public event EventHandler<SpanEventArgs> TraceEnded;
+
+        public Scope Active
+        {
+            get
+            {
+                var activeScope = HttpContext.Current?.Items[_name] as Scope;
+                if (activeScope != null)
+                {
+                    return activeScope;
+                }
+
+                return _activeScopeFallback.Get();
+            }
+        }
+
+        public Scope Activate(Span span, bool finishOnClose)
+        {
+            var newParent = Active;
+            var scope = new Scope(newParent, span, this, finishOnClose);
+            var scopeOpenedArgs = new SpanEventArgs(span);
+
+            SpanOpened?.Invoke(this, scopeOpenedArgs);
+
+            SetScope(scope);
+
+            if (scope.Parent != null)
+            {
+                SpanDeactivated?.Invoke(this, new SpanEventArgs(newParent.Span));
+            }
+
+            SpanActivated?.Invoke(this, scopeOpenedArgs);
+
+            return scope;
+        }
+
+        public void Close(Scope scope)
+        {
+            var current = Active;
+            var isRootSpan = scope.Parent == null;
+
+            if (current != null && current == scope)
+            {
+                // if the scope that was just closed was the active scope,
+                // set its parent as the new active scope
+                SetScope(current.Parent);
+                SpanDeactivated?.Invoke(this, new SpanEventArgs(current.Span));
+
+                if (!isRootSpan)
+                {
+                    SpanActivated?.Invoke(this, new SpanEventArgs(current.Parent.Span));
+                }
+            }
+
+            if (current != null)
+            {
+                SpanClosed?.Invoke(this, new SpanEventArgs(scope.Span));
+
+                if (isRootSpan)
+                {
+                    TraceEnded?.Invoke(this, new SpanEventArgs(scope.Span));
+                }
+            }
+        }
+
+        private void SetScope(Scope scope)
+        {
+            var httpContext = HttpContext.Current;
+            if (httpContext != null)
+            {
+                httpContext.Items[_name] = scope;
+            }
+
+            _activeScopeFallback.Set(scope);
+        }
+    }
+}
+
+#endif
