+++ src/Datadog.Trace.ClrProfiler.Managed/Integrations/AspNetWebApi2Integration.cs
